LayoutTests/fast/exclusions/wrap-parsing-expected.txt

Test parsing of the CSS wrap property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-wrap: auto") is "auto"
PASS test("-webkit-wrap: start 1px") is "start 1px"
PASS test("-webkit-wrap: end 1px 2px") is "end 1px 2px"
PASS test("-webkit-wrap: 5px both 10pt;") is "both 5px 10pt"
PASS test("-webkit-wrap: 2px 3px clear;") is "clear 2px 3px"
PASS test("-webkit-wrap: 5px maximum;") is "maximum 5px"
PASS test("-webkit-wrap: 5px;") is "5px"
PASS test("-webkit-wrap: 5px 10px;") is "5px 10px"
PASS test("-webkit-wrap: none 10px 10em;") is ""
PASS test("-webkit-wrap: auto -5px;") is ""
PASS test("-webkit-wrap: auto both;") is ""
PASS test("-webkit-wrap: auto 1px -10px;") is ""
PASS test("-webkit-wrap: -5px start;") is ""
PASS test("-webkit-wrap: 5px end 'string';") is ""
PASS testComputedStyle("auto") is "auto 0px 0px"
PASS testComputedStyle("start 1px") is "start 1px 0px"
PASS testComputedStyle("end 1px 2px") is "end 1px 2px"
PASS testComputedStyle("5px maximum") is "maximum 5px 0px"
PASS testComputedStyle("5px") is "auto 5px 0px"
PASS testComputedStyle("5px 10px") is "auto 5px 10px"
PASS testComputedStyle("none 5px 10px") is "auto 0px 0px"
PASS testComputedStyle("auto -5px") is "auto 0px 0px"
PASS successfullyParsed is true

TEST COMPLETE
